[FIX] Fixed serialization problem of JobTask [FIX] Naming convention errors [DBG] Debugging tests to put workers processing jobs
Currently I'm facing a problem on the Worker Unit Test, the worker when called to process a job, should have the onlineWorkers with at least one instance, the worker itself.
The workers are being initialized with the workers available on puppet master, but for some reason it seems that .NET Remoting is creating an empty worker from scratch.

[ !!!!! ATENTION !!!!! ] I believe this is the problem, how to fix it?! Am I doing something wrong, help appreciated!